requirements.txt
Pillow >= 10.*.*
mutagen
scandirRecursive @ git+https://github.com/Alejandro-Roldan/scandirRecursive.git
